Added FlowControl enum to step
This should provide the necessary differentiation between manual and automatic tasks.

I think that using an enum is preferable than `bool automatic` for instance, as it can be more easily expanded in the future, if there were tasks with different permission levels for instance which required different steps to be performed by the user.

In terms of laying out a workflow, this could be recorded in JSON or similar as a string value, and then later parsed into the correct enum value.