Added async step executor
This will take a step (and a number of other parameters) and run the step's task, returning an ExecutionResult in a message queue.

By separating out this execution step we can keep the cursor in the main thread only.

I was originally planning to keep this in the .cpp file only, as it is an implementation detail, but clearly there is a need to test it and make sure it is working how I think it is working. Therefore, a definition has been added to the header file and I will create some unit tests for the functionality.