Changed asnc queue to use ResultMessage struct rather than pair
This changes stems from the CPP core guidelines suggestion (I.4) to make interface strongly typed and precise.
In this case, `.id` and `.result` are clearer than `.first` and `second` and more in line with the rest of the code.

It has the added advantage that the message result could be expanded in the future, for example to include a "type", if the message could contain different kinds of information.